Started Test refacturing
