+++ catroidTest/src/org/catrobat/catroid/uitest/ui/activity/SettingsActivityTest.java
@@ -56,124 +56,14 @@ public class SettingsActivityTest extends BaseActivityInstrumentationTestCase<Ma
 		TestUtils.createEmptyProject();
 		settings = solo.getString(R.string.settings);
 		preferences = PreferenceManager.getDefaultSharedPreferences(getActivity());
-		preferences.edit().putBoolean(SettingsActivity.SETTINGS_SHOW_PARROT_AR_DRONE_BRICKS, false).commit();
 		preferences.edit().putBoolean(SettingsActivity.SETTINGS_MINDSTORMS_NXT_BRICKS_ENABLED, false).commit();
 	}
 
 	@Override
 	protected void tearDown() throws Exception {
-		preferences.edit().putBoolean(SettingsActivity.SETTINGS_SHOW_PARROT_AR_DRONE_BRICKS, false).commit();
 		preferences.edit().putBoolean(SettingsActivity.SETTINGS_MINDSTORMS_NXT_BRICKS_ENABLED, false).commit();
 		solo.finishOpenedActivities();
 		super.tearDown();
-	}
-
-	public void testDroneTermsOfUsePermanentAgree() {
-		SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(getActivity());
-
-		preferences.edit().putBoolean(SettingsActivity.SETTINGS_PARROT_AR_DRONE_CATROBAT_TERMS_OF_SERVICE_ACCEPTED_PERMANENTLY, false)
-				.commit();
-
-		assertFalse("Terms of servie should not be accepted",
-				SettingsActivity.areTermsOfServiceAgreedPermanently(getActivity()));
-
-		assertFalse("Terms of servie should not be accepted",
-				preferences.getBoolean(SettingsActivity.SETTINGS_PARROT_AR_DRONE_CATROBAT_TERMS_OF_SERVICE_ACCEPTED_PERMANENTLY, true));
-
-		SettingsActivity.setTermsOfServiceAgreedPermanently(getActivity(), true);
-		assertTrue("Terms of servie should be permanently accepted",
-				SettingsActivity.areTermsOfServiceAgreedPermanently(getActivity()));
-
-		assertTrue("Terms of servie should be permanently accepted",
-				preferences.getBoolean(SettingsActivity.SETTINGS_PARROT_AR_DRONE_CATROBAT_TERMS_OF_SERVICE_ACCEPTED_PERMANENTLY, false));
-	}
-
-	public void testToggleDroneSensors() {
-
-		String forumularEditorDroneSensors = "drone_";
-
-		DroneTestUtils.createStandardDroneProject();
-
-		String dronePreferenceString = solo.getString(R.string.preference_description_quadcopter_bricks);
-
-		solo.waitForActivity(MainMenuActivity.class);
-		UiTestUtils.getIntoScriptActivityFromMainMenu(solo);
-		solo.waitForActivity(ScriptActivity.class);
-		solo.clickOnView(solo.getView(R.id.brick_drone_move_edit_text_second));
-		solo.clickOnText(solo.getString(R.string.formula_editor_sensors));
-		assertFalse("Drone sensors are showing!", solo.searchText(forumularEditorDroneSensors));
-		solo.goBack();
-		solo.waitForActivity(ScriptActivity.class);
-		solo.clickOnActionBarHomeButton();
-		solo.waitForActivity(MainMenuActivity.class.getSimpleName());
-
-		solo.clickOnMenuItem(settings);
-		solo.waitForActivity(SettingsActivity.class.getSimpleName());
-
-		assertTrue("Wrong title", solo.searchText(solo.getString(R.string.preference_title)));
-
-		solo.clickOnText(dronePreferenceString);
-		solo.waitForText(dronePreferenceString);
-		solo.clickOnText(dronePreferenceString);
-		solo.goBack();
-		solo.goBack();
-
-		assertTrue("Drone preference should now be enabled",
-				preferences.getBoolean(SettingsActivity.SETTINGS_SHOW_PARROT_AR_DRONE_BRICKS, false));
-
-		solo.waitForActivity(MainMenuActivity.class);
-		UiTestUtils.getIntoScriptActivityFromMainMenu(solo);
-		solo.waitForActivity(ScriptActivity.class);
-		solo.clickOnView(solo.getView(R.id.brick_drone_move_edit_text_second));
-		solo.clickOnText(solo.getString(R.string.formula_editor_sensors));
-		assertTrue("Drone sensors are not showing!", solo.searchText(forumularEditorDroneSensors));
-	}
-
-	public void testToggleDroneBricks() {
-		String dronePreferenceString = solo.getString(R.string.preference_description_quadcopter_bricks);
-		String categoryDroneLabel = solo.getString(R.string.category_drone);
-		SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(getActivity());
-
-		solo.waitForActivity(MainMenuActivity.class);
-		UiTestUtils.getIntoScriptActivityFromMainMenu(solo);
-		solo.waitForActivity(ScriptActivity.class);
-		UiTestUtils.clickOnBottomBar(solo, R.id.button_add);
-		solo.waitForText(solo.getString(R.string.category_control));
-
-		ListView fragmentListView = solo.getCurrentViews(ListView.class).get(
-				solo.getCurrentViews(ListView.class).size() - 1);
-		solo.scrollListToBottom(fragmentListView);
-
-		assertFalse("Drone brick category is showing!", solo.searchText(categoryDroneLabel));
-		solo.goBack();
-		solo.waitForActivity(ScriptActivity.class);
-		solo.clickOnActionBarHomeButton();
-		solo.waitForActivity(MainMenuActivity.class.getSimpleName());
-
-		solo.clickOnMenuItem(settings);
-		solo.waitForActivity(SettingsActivity.class.getSimpleName());
-
-		assertTrue("Wrong title", solo.searchText(solo.getString(R.string.preference_title)));
-
-		solo.clickOnText(dronePreferenceString);
-		solo.waitForText(dronePreferenceString);
-		solo.clickOnText(dronePreferenceString);
-		solo.goBack();
-		solo.goBack();
-
-		assertTrue("Drone preference should now be enabled",
-				preferences.getBoolean(SettingsActivity.SETTINGS_SHOW_PARROT_AR_DRONE_BRICKS, false));
-
-		solo.waitForActivity(MainMenuActivity.class);
-		UiTestUtils.getIntoScriptActivityFromMainMenu(solo);
-		UiTestUtils.clickOnBottomBar(solo, R.id.button_add);
-
-		solo.waitForText(solo.getString(R.string.category_control));
-
-		fragmentListView = solo.getCurrentViews(ListView.class).get(solo.getCurrentViews(ListView.class).size() - 1);
-		solo.scrollListToBottom(fragmentListView);
-
-		assertTrue("Drone brick category is not showing!", solo.searchText(categoryDroneLabel));
 	}
 
 	public void testToggleMindstormsNXTBricks() {
@@ -208,113 +98,5 @@ public class SettingsActivityTest extends BaseActivityInstrumentationTestCase<Ma
 		solo.scrollListToBottom(fragmentListView);
 		solo.sleep(100);
 		assertTrue("Lego brick category is not showing!", solo.searchText(categoryLegoNXTLabel));
-	}
-
-	public void testOrientation() throws NameNotFoundException {
-		solo.waitForActivity(MainMenuActivity.class);
-		solo.clickOnMenuItem(settings);
-		solo.waitForActivity(SettingsActivity.class.getSimpleName());
-
-		/// Method 1: Assert it is currently in portrait mode.
-		assertEquals("SettingsActivity not in Portrait mode!", Configuration.ORIENTATION_PORTRAIT, getActivity()
-				.getResources().getConfiguration().orientation);
-
-		/// Method 2: Retreive info about Activity as collected from AndroidManifest.xml
-		// https://developer.android.com/reference/android/content/pm/ActivityInfo.html
-		PackageManager packageManager = getActivity().getPackageManager();
-		ActivityInfo activityInfo = packageManager.getActivityInfo(getActivity().getComponentName(),
-				PackageManager.GET_ACTIVITIES);
-
-		// Note that the activity is _indeed_ rotated on your device/emulator!
-		// Robotium can _force_ the activity to be in landscape mode (and so could we, programmatically)
-		solo.setActivityOrientation(Solo.LANDSCAPE);
-
-		assertEquals(
-				SettingsActivity.class.getSimpleName() + " not set to be in portrait mode in AndroidManifest.xml!",
-				ActivityInfo.SCREEN_ORIENTATION_PORTRAIT, activityInfo.screenOrientation);
-	}
-
-	public void testDroneSettingsActivity() {
-
-		solo.clickOnText(solo.getString(R.string.main_menu_continue));
-		solo.clickOnMenuItem(settings);
-		solo.assertCurrentActivity("Wrong Activity", SettingsActivity.class);
-		solo.clickOnText(solo.getString(R.string.preference_description_quadcopter_bricks));
-
-		solo.clickOnText(solo.getString(R.string.brick_drone_set_config));
-		solo.clickOnText(solo.getString(R.string.drone_config_indoor_without_hull));
-		assertTrue("Wrong Settings set, altitude limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_altitude_max_3m)));
-		assertTrue("Wrong Settings set, vertical speed limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_vertical_indoor)));
-		assertTrue("Wrong Settings set, rotation speed limit should be 3 meter", solo.searchText(solo.getString(R
-				.string.drone_set_rotation_indoor)));
-		assertTrue("Wrong Settings set, tilt limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_tilt_indoor)));
-
-		solo.clickOnText(solo.getString(R.string.brick_drone_set_config));
-		solo.clickOnText(solo.getString(R.string.drone_config_outdoor_without_hull));
-		assertTrue("Wrong Settings set, altitude limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_altitude_max_3m)));
-		assertTrue("Wrong Settings set, vertical speed limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_vertical_outdoor)));
-		assertTrue("Wrong Settings set, rotation speed limit should be 3 meter", solo.searchText(solo.getString(R
-				.string.drone_set_rotation_outdoor)));
-		assertTrue("Wrong Settings set, tilt limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_tilt_outdoor)));
-
-		solo.clickOnText(solo.getString(R.string.brick_drone_set_config));
-		solo.clickOnText(solo.getString(R.string.drone_config_indoor_with_hull));
-		assertTrue("Wrong Settings set, altitude limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_altitude_max_3m)));
-		assertTrue("Wrong Settings set, vertical speed limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_vertical_indoor)));
-		assertTrue("Wrong Settings set, rotation speed limit should be 3 meter", solo.searchText(solo.getString(R
-				.string.drone_set_rotation_indoor)));
-		assertTrue("Wrong Settings set, tilt limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_tilt_indoor)));
-
-		solo.clickOnText(solo.getString(R.string.brick_drone_set_config));
-		solo.clickOnText(solo.getString(R.string.drone_config_outdoor_with_hull));
-		assertTrue("Wrong Settings set, altitude limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_altitude_max_3m)));
-		assertTrue("Wrong Settings set, vertical speed limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_vertical_outdoor)));
-		assertTrue("Wrong Settings set, rotation speed limit should be 3 meter", solo.searchText(solo.getString(R
-				.string.drone_set_rotation_outdoor)));
-		assertTrue("Wrong Settings set, tilt limit should be 3 meter", solo.searchText(solo.getString(R.string
-				.drone_set_tilt_outdoor)));
-
-		solo.goBack();
-		solo.goBack();
-	}
-
-	public void testDroneBrickTest() {
-		solo.clickOnText(solo.getString(R.string.main_menu_continue));
-		solo.clickOnMenuItem(settings);
-		solo.assertCurrentActivity("Wrong Activity", SettingsActivity.class);
-		solo.clickOnText(solo.getString(R.string.preference_description_quadcopter_bricks));
-		solo.clickOnText(solo.getString(R.string.preference_description_quadcopter_bricks));
-		solo.goBack();
-		solo.goBack();
-
-		solo.clickOnText(solo.getString(R.string.background));
-		solo.clickOnText(solo.getString(R.string.scripts));
-
-		UiTestUtils.addNewBrick(solo, R.string.brick_drone_takeoff_land);
-		solo.sleep(500);
-		UiTestUtils.dragFloatingBrick(solo, -1.25f);
-
-		UiTestUtils.clickOnPlayButton(solo);
-		assertTrue("DroneBrick present but no drone connection dialog", solo.searchText(solo.getString(R.string.error_no_drone_connected_title)));
-		solo.clickOnText(solo.getString(R.string.close));
-
-		solo.clickOnText(solo.getString(R.string.brick_drone_takeoff_land));
-		solo.clickOnText(solo.getString(R.string.delete));
-		solo.clickOnText(solo.getString(R.string.yes));
-
-		UiTestUtils.clickOnPlayButton(solo);
-
-		assertTrue("DroneBrick present but no drone connection dialog", !solo.searchText(solo.getString(R.string.error_no_drone_connected_title)));
 	}
 }
